Move STATUS and prefix to constants
