+++ src/constants.js
@@ -1,0 +1,13 @@
+const STATUS = {
+  IDLE: 'IDLE',
+  PENDING: 'PENDING',
+  FULFILLED: 'FULFILLED',
+  REJECTED: 'REJECTED',
+}
+
+const prefix = '@@redux-async'
+
+export {
+  STATUS,
+  prefix,
+}
